src/polarization_analysis.py
# Script to analyze polarization of light on epithelial cells
# Polarization Analysis Module for Evolved Epithelial Modulation Detection

import cv2
import numpy as np
import matplotlib.pyplot as plt

class PolarizationAnalyzer:
    def __init__(self):
        pass

    def calculate_polarization(self, image, angle_step=1):
        '''
        Calculate the degree of polarization in the input image across different angles.
        
        Args:
            image (np.array): Input image (polarized).
            angle_step (int): Step size for rotating polarization angles (default is 1 degree).
        
        Returns:
            np.array: Polarization map showing degrees of polarization at each pixel.
        '''
        height, width = image.shape[:2]
        polarization_map = np.zeros((height, width), dtype=np.float32)
        
        # Simulate polarization calculation (placeholder logic)
        for angle in range(0, 180, angle_step):
            rotated_image = self.rotate_polarization(image, angle)
            polarization_map += cv2.absdiff(image, rotated_image)
        
        # Normalize the polarization map
        polarization_map = cv2.normalize(polarization_map, None, 0, 255, cv2.NORM_MINMAX)
        return polarization_map.astype(np.uint8)

    def rotate_polarization(self, image, angle):
        '''
        Rotate the input image to simulate polarization at a specific angle.
        
        Args:
            image (np.array): Input polarized image.
            angle (float): Angle to rotate the polarization.
        
        Returns:
            np.array: Image rotated by the specified polarization angle.
        '''
        h, w = image.shape[:2]
        center = (w // 2, h // 2)
        rotation_matrix = cv2.getRotationMatrix2D(center, angle, 1.0)
        rotated_image = cv2.warpAffine(image, rotation_matrix, (w, h))
        return rotated_image

    def analyze_degree_of_polarization(self, image, threshold=50):
        '''
        Analyze the degree of polarization in the image and highlight regions exceeding a threshold.
        
        Args:
            image (np.array): Input image (polarized).
            threshold (int): Threshold value for detecting regions of strong polarization.
        
        Returns:
            np.array: Binary map showing regions with high degrees of polarization.
        '''
        polarization_map = self.calculate_polarization(image)
        
        # Apply a threshold to the polarization map
        _, high_polarization_map = cv2.threshold(polarization_map, threshold, 255, cv2.THRESH_BINARY)
        
        return high_polarization_map

    def visualize_polarization(self, polarization_map):
        '''
        Visualize the polarization map using a color heatmap.
        
        Args:
            polarization_map (np.array): Input polarization map.
        
        Returns:
            None: Displays the visualization.
        '''
        plt.imshow(polarization_map, cmap='hot')
        plt.colorbar()
        plt.title('Degree of Polarization Heatmap')
        plt.show()

    def detect_polarization_anomalies(self, image, reference_image, threshold=50):
        '''
        Detect anomalies based on polarization differences between two images.
        
        Args:
            image (np.array): Input polarized image.
            reference_image (np.array): Reference polarized image for comparison.
            threshold (int): Threshold for identifying significant polarization differences.
        
        Returns:
            np.array: Anomaly map showing regions with significant polarization differences.
        '''
        # Step 1: Compute polarization maps for both images
        polarization_map1 = self.calculate_polarization(image)
        polarization_map2 = self.calculate_polarization(reference_image)
        
        # Step 2: Compute the absolute difference between the two polarization maps
        polarization_difference = cv2.absdiff(polarization_map1, polarization_map2)
        
        # Step 3: Apply a threshold to isolate significant differences
        _, anomaly_map = cv2.threshold(polarization_difference, threshold, 255, cv2.THRESH_BINARY)
        
        return anomaly_map

    def compare_polarization_across_channels(self, image):
        '''
        Compare polarization differences across different color channels (RGB).
        
        Args:
            image (np.array): Input image (polarized).
        
        Returns:
            dict: A dictionary containing polarization maps for each color channel (R, G, B).
        '''
        # Split the image into its color channels
        channels = cv2.split(image)
        polarization_maps = {}

        # Calculate polarization for each channel
        for i, channel in enumerate(['R', 'G', 'B']):
            polarization_maps[channel] = self.calculate_polarization(channels[i])

        return polarization_maps

    def highlight_polarization_anomalies(self, image, anomaly_map, color=(255, 0, 0)):
        '''
        Highlight polarization anomalies on the original image.
        
        Args:
            image (np.array): Input polarized image.
            anomaly_map (np.array): Binary anomaly map.
            color (tuple): Color to use for highlighting anomalies (default is red).
        
        Returns:
            np.array: Image with highlighted polarization anomalies.
        '''
        highlighted_image = image.copy()
        highlighted_image[anomaly_map > 0] = color
        return highlighted_image
